Fix building on Linux

+++ source/ResourceFile.cpp
@@ -58,8 +58,8 @@ namespace HAYDEN
         entryStart = seekDistance;
 
         // Get decompressed & compressed sizes relative to entryStart.
-        lwoHeader.decompressedSize = *(int*)(lwoDecompressedHeader.data() + (lwoDecompressedHeader.size() - (entryStart + 8i64)));
-        lwoHeader.compressedSize = *(int*)(lwoDecompressedHeader.data() + (lwoDecompressedHeader.size() - (entryStart + 4i64)));
+        lwoHeader.decompressedSize = *(int*)(lwoDecompressedHeader.data() + (lwoDecompressedHeader.size() - (size_t)(entryStart + 8)));
+        lwoHeader.compressedSize = *(int*)(lwoDecompressedHeader.data() + (lwoDecompressedHeader.size() - (size_t)(entryStart + 4)));
         return lwoHeader;
     }
     std::vector<byte> ResourceFile::GetEmbeddedFileHeader(FILE* f, const uint64 fileOffset, const uint64 compressedSize) const
